+++ wearcapture/__init__.py
@@ -5,4 +5,5 @@ __all__ = [
     "adb",
     "capture_engine",
     "image_ops",
+    "profiles",
 ]